fix: address integration layout review feedback
